Util.TreeGenerator

Description

Module used for generate 'NTree Cell' from lists.

Documentation

type Name = String

Name of a level (= name of a column = label name).

type Level = (Name, [[[String]]])

Level, represent a level of the tree in a list way. There is a list in the first level of the list, for each nodes on the level above. And the list in the second level contains lists representing nodes for this Level.

generate

Arguments

:: [Level]
List of levels.

-> NTree Cell
The tree built from the levels.

generate, generate a tree given the levels.